+++ python/pyspark/pandas/tests/indexes/test_datetime_property.py
@@ -17,45 +17,11 @@
 import numpy as np
 import pandas as pd
 
-import pyspark.pandas as ps
-from pyspark.testing.pandasutils import PandasOnSparkTestCase, TestUtils
+from pyspark.testing.pandasutils import PandasOnSparkTestCase
+from pyspark.pandas.tests.indexes.test_datetime import DatetimeIndexTestingFuncMixin
 
 
-class DatetimeIndexPropertyTestsMixin:
-    @property
-    def fixed_freqs(self):
-        return [
-            "D",
-            "H",
-            "T",  # min
-            "S",
-            "L",  # ms
-            "U",  # us
-            # 'N' not supported
-        ]
-
-    @property
-    def non_fixed_freqs(self):
-        return ["W", "Q"]
-
-    @property
-    def pidxs(self):
-        return [
-            pd.DatetimeIndex([0]),
-            pd.DatetimeIndex(["2004-01-01", "2002-12-31", "2000-04-01"]),
-        ] + [
-            pd.date_range("2000-01-01", periods=3, freq=freq)
-            for freq in (self.fixed_freqs + self.non_fixed_freqs)
-        ]
-
-    @property
-    def psidxs(self):
-        return [ps.from_pandas(pidx) for pidx in self.pidxs]
-
-    @property
-    def idx_pairs(self):
-        return list(zip(self.psidxs, self.pidxs))
-
+class DatetimeIndexPropertyTestsMixin(DatetimeIndexTestingFuncMixin):
     def test_properties(self):
         for psidx, pidx in self.idx_pairs:
             self.assert_eq(psidx.year, pidx.year)
@@ -83,7 +49,10 @@ class DatetimeIndexPropertyTestsMixin:
             self.assert_eq(psidx.isocalendar().week, pidx.isocalendar().week.astype(np.int64))
 
 
-class DatetimeIndexPropertyTests(DatetimeIndexPropertyTestsMixin, PandasOnSparkTestCase, TestUtils):
+class DatetimeIndexPropertyTests(
+    DatetimeIndexPropertyTestsMixin,
+    PandasOnSparkTestCase,
+):
     pass
